mmc: give SDR50 precedence over DDR50
The MMC driver prefers DDR50 over SDR50 when both are available.  The
reasoning is that DDR50 uses a 50MHz bus clock and therefore should be
more reliable.  This isn't actually true in practice, thanks to the
rather tight setup and hold requirements of double data rate clocking.
When living with a cellular radio, 100MHz has fewer problematic harmonics
than 50MHz does, so use SDR50 if it is supported.

Also, throttle the bus on errors based on the current speed and not the
card capabilities.  This avoids odd situations where the host doesn't
support a speed that the card does.

Change-Id: I0bbd2e7bdc230675502a46dc7e2e7b4107f88605
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.mot.com/719500
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>